% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv.methyl-package.R
\docType{package}
\name{cnv.methyl-package}
\alias{cnv.methyl}
\alias{cnv.methyl-package}
\title{cnv.methyl: cnv.methyl finds copy number alteration from methylation idat files}
\description{
The package implements an automatic and dynamic thresholding algorithm to find somatic copy number alterations in methylation arrays based on log2R intensities. It relies mostly on minfi for preprocessing and and conumee for segmentation.
}
\author{
\strong{Maintainer}: Izar Llaquet de Villasante \email{idevillasante@carrerasresearch.org} (\href{https://orcid.org/0000-0001-7325-0631 }{ORCID})

}
\keyword{internal}
