% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kc_get.R
\name{Kc_get}
\alias{Kc_get}
\title{Set of functions to obtain Kc from reference copy number state.}
\usage{
Kc_get(
  ID,
  ss,
  ref_genes = "all",
  segfile_folder = "analysis/CONUMEE/Segments",
  log2rfile_folder = "analysis/CONUMEE/log2r",
  arraytype = "450K",
  anno_file = NULL,
  cn_genes = NULL,
  Kc_method = "balanced"
)
}
\arguments{
\item{ID}{ID inside Sample sheet and in the Sample_Name.}

\item{ss}{sample sheet}

\item{ref_genes}{Granges object with subset of genes to use for cnv analysis.}

\item{arraytype}{Methylation array type}

\item{anno_file}{anno file CNV.anno object saved as .rds. default
CNV_Germline_GSITIC2_BROAD_SNP6.merged.151117.hg19.CNV.txt}

\item{cn_genes}{Used to subset genes from the reference set. Needed in order
to generate new Kc from a reference set, where cn_genes are previously known
genes to be altered in a given cn_state. Otherwise use ref_genes.}

\item{Kc_method}{Choose which Kc you want to use. Either "curated" or "balanced.
The first comes from a list of 94 genes manually selected by experts and an
uneven proportion of cancers. The second one comes from all genes and balanced
proportion of cancer types. default= balanced}
}
\value{
GRanges object of the genes of interest "ref_genes" and a cna column
with the copy number alterations for each of them.
}
\description{
Set of functions to obtain Kc from reference copy number state.
}
\examples{
#data("TrainingSet_Sample_sheet")
#ss<-TrainingSet_Sample_sheet[60:200,]
ID="TCGA-19-A6J4-01A-11D-A33U-05"
ss[Sample_Name==ID,]->ss
cna<-Kc_get(ss=ss,ID="TCGA-19-A6J4-01A-11D-A33U-05")
cna
}
\author{
Izar de Villasante
}
