% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\alias{pre_process.myLoad}
\alias{qc}
\alias{purify}
\alias{queryfy}
\title{Pre-process methylation array data from (raw) .idat files to segment & log2r intensity files calculated by CONUMEE.}
\usage{
pre_process(
  targets,
  purity = NULL,
  query = T,
  RGset = T,
  out = "./analysis/intermediate/",
  idats_folder = NULL,
  subf = "IDATS/",
  folder = NULL,
  ncores = NULL,
  arraytype = NULL,
  copy = FALSE,
  frac = 0.1,
  pval = 0.01,
  remove_sex = TRUE,
  sampGroups = NULL
)

pre_process.myLoad(
  targets,
  folder,
  arraytype = "450K",
  ncores = NULL,
  files = NULL,
  copy = FALSE
)

qc(
  rgSet,
  sampGroups = NULL,
  sampNames = "Sample_Name",
  qc_folder = "analysis/intermediate/QC/"
)

purify(myLoad, knn = 5)

queryfy(
  targets,
  rgSet,
  sampGroups = NULL,
  sampNames = "Sample_Name",
  frac = 0.1,
  pval = 0.01,
  remove_sex = TRUE,
  arraytype = NULL,
  qc_folder = "analysis/intermediate/QC"
)
}
\arguments{
\item{targets}{A sample sheet with the required fields for ChAMP to load files.}

\item{purity}{Whether or not you want purity imputed by Rfpurify.
Default=TRUE.}

\item{query}{Set to TRUE when you are only interested in imputing purity.
Default=TRUE}

\item{RGset}{Whether you want normalised "RGChannelSet" or
"RGChannelSetExtended" to be saved or not. path=out}

\item{out}{Results folder to store the normalized and clean RGset.}

\item{idats_folder}{Folder containing idats. It will try to guess which idat
belongs to which row in the targets data.frame.}

\item{subf}{Subfolder inside results folder where IDAT files will be copied.}

\item{folder}{this directory will be created as the combination of out
and subf. if you save the csv file inside this folder it will be used by
minfi::read.450k.sheet.}

\item{ncores}{number of cores to use.}

\item{arraytype}{Methylation array type}

\item{copy}{Change to TRUE if you want to copy files to idats folder.
Useful when working in HPC and idats are in ISILON.}

\item{frac}{Fraction of faulty probes (P-value > pval ) allowed. default 0.1}

\item{pval}{P-value cutoff. default 0.01}

\item{remove_sex}{Boolean. Should sex chromosomes be removed? default = TRUE}

\item{sampGroups}{Groups for qc report & plots color category.}

\item{files}{Files}

\item{rgSet}{Input RGChannelSet object with raw idats and metadata.}

\item{sampNames}{Sample names to be used for labels.}

\item{qc_folder}{Path to location where qc_report and plot will be saved.}

\item{myLoad}{a RGset as returned by minfi::read.metharray.exp. will use this as basedir to
load the idats. default is results/IDATS/.}

\item{knn}{number of neighbors for knn algorithm}
}
\value{
Log2r intensities ready to be analysed with conumee.

RGset

Qc plots

Normalized & filtered RGset.
}
\description{
construct RGChannelSet in parallel using foreach

Generate Qc reports
}
\examples{

data("TrainingSet_Sample_sheet")
ss<-TrainingSet_Sample_sheet[1:56,]
pre_process(ss)
}
\author{
izar de Villasante

izar de Villasante
}
