% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{read.metharray.exp.par}
\alias{read.metharray.exp.par}
\title{construct RGChannelSet in parallel}
\usage{
read.metharray.exp.par(
  targets,
  folder,
  files,
  copy = FALSE,
  verbose = TRUE,
  arraytype = "450K",
  ncores = NULL,
  extended = FALSE,
  force = FALSE
)
}
\arguments{
\item{targets}{A sample sheet with the required fields for minfi to load files.}

\item{folder}{this directory will be created as the combination of out
and subf. if you save the csv file inside this folder it will be used by
minfi::read.450k.sheet.}

\item{files}{Files}

\item{copy}{Change to TRUE if you want to copy files to idats folder.
Useful when working in HPC and idats are in ISILON.}

\item{verbose}{Should the function be verbose?}

\item{arraytype}{Methylation array type}

\item{ncores}{number of cores to use.}

\item{extended}{Should a \code{RGChannelSet} or a
    \code{RGChannelSetExtended} be returned. }

\item{force}{Should reading different size IDAT files be forced? See
    Details.}
}
\value{
RGset
}
\description{
construct RGChannelSet in parallel using foreach
}
\author{
izar de Villasante
}
