% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cnv.methyl.R
\name{run_cnv.methyl}
\alias{run_cnv.methyl}
\title{Generate segment & log2r intensity files calculated by CONUMEE.}
\usage{
run_cnv.methyl(
  targets,
  ref_genes = "all",
  cn_genes = NULL,
  Kc_method,
  out = "analysis/intermediate/",
  RGset = T,
  purity = NULL,
  arraytype = "450K",
  folder = NULL,
  anno_file = NULL,
  ctrl_file = "WB",
  Sample_Name = NULL,
  ncores = NULL,
  seg.folder = "Segments",
  log2r.folder = "log2r",
  conumee.folder = "analysis/CONUMEE/",
  probeid = "probeid",
  intensities = NULL
)
}
\arguments{
\item{targets}{A sample sheet with the required fields for minfi to load files.}

\item{ref_genes}{Granges object with subset of genes to use for cnv analysis.}

\item{cn_genes}{Used to subset genes from the reference set. Needed in order
to generate new Kc from a reference set, where cn_genes are previously known
genes to be altered in a given cn_state. Otherwise use ref_genes.}

\item{Kc_method}{Choose which Kc you want to use. Either "curated" or "balanced.
The first comes from a list of 94 genes manually selected by experts and an
uneven proportion of cancers. The second one comes from all genes and balanced
proportion of cancer types. default= balanced}

\item{out}{Parent working directory where you want to have your results.}

\item{RGset}{Whether you want normalised "RGChannelSet" or
"RGChannelSetExtended" to be saved or not. path=out}

\item{purity}{Whether or not you want purity imputed by Rfpurify.
Default=TRUE.}

\item{arraytype}{Methylation array type}

\item{folder}{this directory will be created as the combination of out
and subf. if you save the csv file inside this folder it will be used by
minfi::read.450k.sheet.}

\item{anno_file}{anno file CNV.anno object saved as .rds. default
CNV_Germline_GSITIC2_BROAD_SNP6.merged.151117.hg19.CNV.txt}

\item{ctrl_file}{CNV data object with intensities from the controls group.
must be compressed as '.rds' file format, default uses 96 WB samples.
Default='WB'}

\item{Sample_Name}{Samples to be analysed. If is NULL all columns in input
file will be used. Accepts numbered index and names. Default=NULL}

\item{ncores}{number of cores to use.}

\item{seg.folder}{Subfolder inside results where segment files are saved.
The segments are generated with CONUMEE CNV.segment(CNV.detail(CNV.bin(fit)))
default = "Segments"}

\item{log2r.folder}{Subfolder inside results where log2r values are saved.
log2r values are the log2 ratio of the intensities(_GRN + _RED channel) between the
query and the reference set (control) as returned by CNV.fit function from CONUMEE.
default = "log2r"}

\item{conumee.folder}{Parent working directory where you want to have your results.}

\item{probeid}{name of column with probe ids.}

\item{intensities}{dataset with intensities. either dataframe or path to file.
For big datasets '.fst' file is recomended.}
}
\value{
data.frame with metadata from segmentation, genomic ranges, genes
and scna for each of the genes
}
\description{
Generate segment & log2r intensity files calculated by CONUMEE.
}
\examples{
#data("TrainingSet_Sample_sheet")
#ss<-TrainingSet_Sample_sheet[1:56,]
t0<-Sys.time()
cnv<-cnv.methyl(targets=ss)
t1<-Sys.time()
cnv
print(paste0("elapsed time: ",t1-t0))
}
