% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMPextr.R
\name{DMPextr}
\alias{DMPextr}
\title{Extract DMPs, annotation and methylation difference for each contrast}
\usage{
DMPextr(
  fit,
  ContrastsDM,
  p.value,
  beta_normalized,
  mDiff,
  ann,
  writeOut = TRUE
)
}
\arguments{
\item{fit}{list containing a linear model fit produced by \code{lmFit}, \code{lm.series}, \code{gls.series} or \code{mrlm}.
  For \code{topTable}, \code{fit} should be an object of class \code{MArrayLM} as produced by \code{lmFit} and \code{eBayes}.}

\item{ContrastsDM}{list of contrasts as returned by limma::makeContrasts()
which will pass to limma topTable as input}

\item{p.value}{cutoff value for adjusted p-values. Only genes with lower p-values are listed.}

\item{beta_normalized}{normalized betavalues, as produce by minfi::getBeta(grSet_noob)),
where colnames(beta_normalized) == metadata$sample_Name}

\item{mDiff}{absolute mean methylation difference between groups to filter by}

\item{ann}{annotation dataset from manifest with metadata such as gene info,
CGI, RefGene, etc. see topTable genelist arg.}

\item{write.out}{save result as .csv default = TRUE.}
}
\value{
data.table
}
\description{
For each contrast extract a set of DMPs and add gene annotation and methylation values
}
\examples{

betas<-readRDS("data/betas.rds")
fit<-readRDS("data/fit2.rds")
ann<-readRDS("data/ann.rds")
DMPann <- DMPextr(fit = fit,                       # linear contrast model
                  ContrastsDM = ContrastsDM,          # contrasts
                  p.value = 0.01,                      # filter significantly different probes
                  beta_normalized = beta_noob,        # extract mean group betas
                  mDiff = 0.5,                        # select mean methylation differences
                  ann = ann,                          # annotate positions (CGI, RefGene, etc)
                  writeOut = FALSE                    # write output to file
}
\author{
Izar de Villasante

Angelika Merkel
}
