% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kc_make.R
\name{Kc_make}
\alias{Kc_make}
\title{Set of functions to obtain Kc from reference copy number state.}
\usage{
Kc_make(
  ss,
  arraytype = NULL,
  cncols = c("Amp10", "Amp", "Gains", "HetLoss", "HomDel"),
  segfile_folder = "analysis/CONUMEE/Segments",
  anno_file = NULL,
  ncores = NULL,
  log2rfile_folder = "analysis/CONUMEE/log2r",
  ref_genes = "paper"
)
}
\arguments{
\item{ss}{Sample sheet containing purities, cnstate and Sample_Name ids.}

\item{arraytype}{Methylation array type}

\item{cncols}{column names containing reference set of genes for each cn state}

\item{anno_file}{anno file CNV.anno object saved as .rds. default
CNV_Germline_GSITIC2_BROAD_SNP6.merged.151117.hg19.CNV.txt}

\item{ref_genes}{Granges object with subset of genes to use for cnv analysis.
default = "paper"}

\item{fname}{identifier for your Kc reference data used in the output names.}
}
\description{
Set of functions to obtain Kc from reference copy number state.
}
\examples{
library(cnv.methyl)
data("ss")
Kc_make(ss=ss,cncols=c("Amp","Amp10","Gains"))->a
}
