% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_process.R
\name{pre_process}
\alias{pre_process}
\alias{pre_process.myLoad}
\alias{purify}
\alias{queryfy}
\title{Pre-process methylation array data from (raw) .idat files to segment & log2r intensity files calculated by CONUMEE.}
\usage{
pre_process(
  targets,
  purity = NULL,
  query = T,
  RGset = T,
  out = "analysis/intermediate/",
  subf = "IDATS/",
  folder = NULL,
  ncores = NULL,
  arraytype = "450K"
)

pre_process.myLoad(targets, folder, arraytype = "450K", ncores = NULL, ...)

purify(myLoad, knn = 5)

queryfy(myLoad)
}
\arguments{
\item{targets}{A sample sheet with the required fields for ChAMP to load files.}

\item{purity}{Whether or not you want purity imputed by Rfpurify.
Default=TRUE.}

\item{query}{Set to TRUE when you are only interested in imputing purity.
Default=TRUE}

\item{RGset}{Whether you want normalised "RGChannelSet" or
"RGChannelSetExtended" to be saved or not. path=out}

\item{out}{Results folder to store the normalized and clean RGset.
Default = NULL}

\item{subf}{Subfolder inside results folder where IDAT files will be copied.}

\item{folder}{this directory will be created as the combination of out
and subf. if you save the csv file inside this folder it will be used by
minfi::read.450k.sheet.}

\item{ncores}{number of cores to use.}

\item{arraytype}{Methylation array type}

\item{...}{optional arguments to read.metharray.exp}

\item{myLoad}{a RGset as returned by minfi::read.metharray.exp. will use this as basedir to
load the idats. default is results/IDATS/.}

\item{knn}{number of neighbors for knn algorithm}
}
\value{
Log2r intensities ready to be analysed with conumee.

RGset
}
\description{
construct RGChannelSet in parallel using foreach
}
\examples{

data("TrainingSet_Sample_sheet")
ss<-TrainingSet_Sample_sheet[1:56,]
pre_process(ss)
}
\author{
izar de Villasante
}
